+++ app/src/main/java/me/juangoncalves/mentra/db/models/WalletValueModel.kt
@@ -4,6 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Entity(
@@ -23,6 +24,6 @@ import java.time.LocalDate
 )
 data class WalletValueModel(
     @ColumnInfo(name = "wallet_id") val walletId: Long,
-    @ColumnInfo(name = "usd_value") val valueInUSD: Double,
+    @ColumnInfo(name = "usd_value") val valueInUSD: BigDecimal,
     @ColumnInfo(name = "date") val date: LocalDate = LocalDate.now()
 )
